requirements.txt
pymodbus==2.5.3
pyyaml
